requirements.txt
Django
gunicorn
django-allauth
dj-rest-auth
djangorestframework
djangorestframework-simplejwt
python-dotenv
requests
cryptography
mysqlclient
